Module pomelo_hir::hir

Definition of the hir.

This module’s contents are exported at the crate root, this is just separated for some internal organization.

Structs

ConBind

Binding of a datatype constructor.

DataBind

Binding of a new datatype.

Dec

HIR declaration node.

ExpRow

Record entry in an expression.

Expr

HIR expression node.

FloatWrapper

Wrapper so we can derive Eq for Scon.

MRule

A rule in a match statement.

Pat

HIR pattern node.

Ty

HIR type node.

TyRow

Record entry in a type.

TypBind

Binding of a type constructor.

ValBind

Binding of the names in a pattern to an expression.

Enums

DecKind

Kinds of HIR declarations.

DefLoc

Location where an identifier is bound.

ExBind

Binding an exception.

ExprKind

Kinds of HIR expressions.

Fixity

Type of fixity, including whether it is left- or right-associative and its operator precedence.

PatKind

Kinds of HIR expressions.

PatRow

Record entry in a pattern.

Scon

HIR constant (literal).

TyKind

Kinds of HIR types.